// FrontEnd/src/app/services/api-backend/model/clinica.ts
/**
 * Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Sucursale } from './sucursale';
import { Cita } from './cita';


export interface Clinica { 
    idClinica?: number;
    idSucursal?: number | null;
    nombre?: string | null;
    estado?: string | null;
    idSucursalNavigation?: Sucursale;
    cita?: Array<Cita> | null;
}